refactor: extract PackDetailView into focused extensions
Split the 1,100-line PackDetailView.swift into 6 focused files following
the same extension pattern used for LiveKeyboardOverlayView:

- PackDetailView.swift (core): struct, properties, body, header, description, quick settings
- PackDetailView+BindingRows.swift: embedded editors and binding row rendering
- PackDetailView+PackActions.swift: install/uninstall/apply methods
- PackDetailView+LiveState.swift: live state query methods
- PackDetailView+Toast.swift: toast overlay and temporary error display
- PackDetailView+Helpers.swift: display helpers, UndoSnapshot struct
